[Editorial] replace all page references with anchor links
Instead of pages we link to the nearest block element, or to a specific endnote if that’s obviously the target. Not all of these references make sense, but unless it was an obvious mistake with an obvious correction I’ve left them pointing to broadly the same place. The Gutenberg volunteers assumed that some references point into this production when they’re really to other works; those have been removed.